+++ src/plone/app/fhirfield/__init__.py
@@ -3,6 +3,8 @@
 from plone.app.fhirfield import patch  # noqa: I001
 patch.monkey_patch_fhir_base_model()  # noqa: I003
 
+
+from .exc import SearchQueryError  # noqa: I001,F401
 from .field import FhirResource  # noqa: I001,F401
 from .widget import FhirResourceWidget  # noqa: I001,F401